+++ stroom-core-shared/src/main/java/stroom/data/shared/DataResource.java
@@ -16,9 +16,6 @@
 
 package stroom.data.shared;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import org.fusesource.restygwt.client.DirectRestService;
 import stroom.meta.shared.FindMetaCriteria;
 import stroom.pipeline.shared.AbstractFetchDataResult;
 import stroom.util.shared.ResourceGeneration;
@@ -26,10 +23,15 @@ import stroom.util.shared.ResourceKey;
 import stroom.util.shared.ResourcePaths;
 import stroom.util.shared.RestResource;
 
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
+import org.fusesource.restygwt.client.DirectRestService;
+
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
@@ -54,8 +56,12 @@ public interface DataResource extends RestResource, DirectRestService {
     ResourceKey upload(UploadDataRequest request);
 
     @GET
+    @Path("/{streamId}")
+    @ApiOperation(
+            value = "Fetch data",
+            response = AbstractFetchDataResult.class)
     AbstractFetchDataResult fetchData(
-            final @QueryParam("streamId") Long streamId,
+            final @PathParam("streamId") long streamId,
             final @QueryParam("streamsOffset") Long streamsOffset,
             final @QueryParam("streamsLength") Long streamsLength,
             final @QueryParam("pageOffset") Long pageOffset,